image download bug fixed

+++ src/pages/product-detail/product-detail.ts
@@ -106,46 +106,55 @@ export class ProductDetailPage {
   private downloadImagesAndVideo() {
     let slf = this;
     this.photoLibrary.requestAuthorization({read: true, write: true}).then(() => {
-      // download images
-      if (this.currentProduct.imageSrcs && this.currentProduct.imageSrcs.length > 0) {
-        for (let imgSrc of this.currentProduct.imageSrcs) {
-          this.photoLibrary.saveImage(imgSrc, '阿坤珠宝');
-        }
-      }
-
       // copy product code, name and description to clipboard
       slf.clipboard.copy("【编号】：" + slf.currentProduct.code + "\r\n【说明】:" + (slf.currentProduct.name||'') 
         + "\r\n【出厂价】：" + (slf.currentProduct.exFactoryPrice));
 
-      // download video
-      this.fileTransfer.download(this.currentProduct.videoSrc, this.file.tempDirectory + "AKunZhuBao.mp4")
-        .then((entry)=>{
-          this.photoLibrary.saveVideo(entry.toURL(), '阿坤珠宝').then(()=>{
-            // remove file
-            this.file.removeFile(this.file.tempDirectory, 'AKunZhuBao.mp4')
-              .then(()=>console.log('Remove-Video succeeded.'), (err)=> console.log('Remove-Video failed. Caused By : ' + err));
-            this._showMessage("视频已下载到本地相册'阿坤珠宝'");
-            Wechat.jumpToWechat("weixin://", function(){}, function(){ this._showMessage('跳转微信失败');});
-          }, ()=>{
-            this._showMessage('视频已下载失败');
-          });
-        }, (err)=>{
-          console.log("FileTransfer Download failed. Caused by : " + err);
-        }).catch((err)=>{ console.log("FileTransfer Download failed. Caused by : " + err); });
+      // download images
+      let i = 0; 
+      let len = 0;
+      if (this.currentProduct.imageSrcs && this.currentProduct.imageSrcs.length > 0) {
+        len = this.currentProduct.imageSrcs.length;
+        for (; i < len; i++) {
+          this.photoLibrary.saveImage(this.currentProduct.imageSrcs[i], '阿坤珠宝');
+        }
+      }
 
-      let alert = this.alertController.create({
-        title: "提示",
-        subTitle: "图片已下载到相册'阿坤珠宝'",
-        buttons: [{
-          text: 'OK',
-          handler: ()=> {
-            Wechat.jumpToWechat("weixin://", function(){}, function(){
-              this._showMessage('跳转微信失败');
+      // download video
+      if (this.currentProduct.hasVideo) {
+        this.fileTransfer.download(this.currentProduct.videoSrc, this.file.tempDirectory + "AKunZhuBao.mp4")
+          .then((entry)=>{
+            this.photoLibrary.saveVideo(entry.toURL(), '阿坤珠宝').then(()=>{
+              // remove file
+              this.file.removeFile(this.file.tempDirectory, 'AKunZhuBao.mp4')
+                .then(()=>console.log('Remove-Video succeeded.'), (err)=> console.log('Remove-Video failed. Caused By : ' + err));
+              this._showMessage("视频已下载到本地相册'阿坤珠宝'");
+              Wechat.jumpToWechat("weixin://", function(){}, function(){ this._showMessage('跳转微信失败');});
+            }, ()=>{
+              this._showMessage('视频已下载失败');
             });
-          }
-        }]
-      });
-      alert.present();
+          }, (err)=>{
+            console.log("FileTransfer Download failed. Caused by : " + err);
+          }).catch((err)=>{ console.log("FileTransfer Download failed. Caused by : " + err); });
+      }
+
+      if (len > 0 || this.currentProduct.hasVideo) {
+        setTimeout(function(){
+          let alert = slf.alertController.create({
+            title: "提示",
+            subTitle: "图片已下载到相册'阿坤珠宝'",
+            buttons: [{
+              text: 'OK',
+              handler: ()=> {
+                Wechat.jumpToWechat("weixin://", function(){}, function(){
+                  slf._showMessage('跳转微信失败');
+                });
+              }
+            }]
+          });
+          alert.present();
+        }, 1500);
+      }
     }).catch(err => console.log("Save Images&Video failed. Caused By : " + err));
   }
